requirements.txt
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
isodate==0.6.1
lxml==5.1.0
more-itertools==10.3.0
platformdirs==4.2.0
pymtom_xop==0.0.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-file==2.0.0
requests-toolbelt==1.0.0
six==1.16.0
urllib3==2.2.1
zeep==4.2.1
